requirements.txt
dash>=2.2.0
dash-bootstrap-components>=1.4.1
Flask>=2.0.3
gunicorn>=20.1.0
newsapi-python>=0.2.7
nltk>=3.8.1
numpy>=1.24.2
pandas>=1.5.3
plotly>=5.13.1
pytrends>=4.9.2
requests>=2.28.2
